requirements.txt
python-dotenv==1.2.1
langchain-community==0.4.1
langchain-core==1.2.9
langchain-text-splitters==1.1.0
langchain-chroma==1.1.0
gpt4all==2.8.2
unstructured[pdf]==0.18.31
unstructured[md]==0.18.31
langchain-google-genai==4.2.0
pytest=9.0.2